+++ tranquilizer/resource.py
@@ -1,30 +1,41 @@
 '''construct flask-restplus resources'''
-from flask import jsonify, request
+from flask import jsonify
 from flask_restplus import Resource, reqparse
 from collections import Mapping, Sequence
+from typing import List
 
 from .types import is_container, type_mapper
 
 
-def _make_parser(func_spec, location='args'):
+def make_parser(func_spec, location='args', compat=False):
     '''Create RequestParser from annotated function arguments
 
-    arguments without default values are flagged as required'''
+    arguments without default values are flagged as required
+
+    Parameters
+    ----------
+    :param func_spec: The decorated functions spec
+    :param location: expected location of arguments in request
+                     'args', or 'body'
+    :param compat: If compatibility with anaconda-enterprise-web-publisher
+                   is required. (default: False)
+    '''
 
     parser = reqparse.RequestParser(bundle_errors=True)
-    for argument,spec in func_spec['args'].items():
+    for argument, spec in func_spec['args'].items():
+        if compat and spec.get('annotation', None) is None:
+            # un-typed arguments are returned to the
+            # function as a list of strings by @publish
+            action = 'append'
+            _type = List[str]
+        else:
+            # @tranquilize assumes untyped
+            # arguments are strings
+            _type = spec.get('annotation', str)
+            action = 'store'
 
-        _type = spec.get('annotation', str)
         _default = spec.get('default', None)
-        action = 'store'
-
         _type = type_mapper(_type)
-
-#       try:
-#           description = getattr(_type, '__description__')
-#       except AttributeError:
-#           description = None
-
         doc = func_spec['param_docs'].get(argument, None)
 
         # Files (e.g., images) arrive in a different
@@ -36,15 +47,12 @@ def _make_parser(func_spec, location='args'):
         except AttributeError:
             _location = location
 
-
         if is_container(_type):
             action = 'append'
-            type_name = _type.__args__[0].__name__
-            _type.__schema__ = {'type':type_name}
 
         parser.add_argument(argument, type=_type,
                             default=_default,
-                            required=(not 'default' in spec),
+                            required=(not ('default' in spec)),
                             location=_location,
                             action=action,
                             help=doc)
@@ -53,22 +61,41 @@ def _make_parser(func_spec, location='args'):
 
 
 def make_resource(func, api):
-    location = 'form' if func._method in ['put','post'] else 'args'
-    parser = _make_parser(func._spec, location=location)
+    if func._methods:
+        return _make_resources(func, api)
+    else:
+        return _make_resource(func, api, func._method)
+
+
+def _make_resources(func, api):
+    '''Provide compatibility with web-publisher'''
+    resources = {}
+    for m in func._methods:
+        resources[m] = getattr(_make_resource(func, api, m), m)
+
+    Tranquilized = type('Tranquilized', (Resource,), resources)
+
+    return Tranquilized
+
+
+def _make_resource(func, api, method):
+    location = 'form' if method in ['put','post'] else 'args'
+    compat = True if (func._methods is not None) else False
+    parser = make_parser(func._spec, location=location, compat=compat)
 
     @api.expect(parser, validate=True, strict=True)
     def _method(self):
-        request = parser.parse_args()
-        output = func(**request)
+        req = parser.parse_args()
+        output = func(**req)
         return jsonify(output)
 
     error_docs = func._spec['error_docs']
     if error_docs:
-        _method = api.doc(responses = error_docs)(_method)
+        _method = api.doc(responses=error_docs)(_method)
 
     _method.__doc__ = func._spec['docstring']
 
-    Tranquilized = type('Tranquilized', (Resource,), {func._method:_method})
+    Tranquilized = type('Tranquilized', (Resource,), {method:_method})
 
     return Tranquilized