+++ tranquilizer/types.py
@@ -1,22 +1,35 @@
-from collections import Mapping, Sequence
+from collections.abc import Mapping, Sequence
 from dateutil.parser import parse
 from datetime import datetime, date
-from typing import List, Generic, TypeVar, TextIO, BinaryIO
+from typing import TextIO, BinaryIO
 from werkzeug.datastructures import FileStorage
+from flask_restplus.reqparse import PY_TYPES
+from flask_restplus import fields
 import io
-
-T = TypeVar('T')
-S = TypeVar('S')
 
 def is_container(type_):
     '''Test if provided type function is a scalar
 
     strings and bytes are considered to be scalars.'''
 
+    origin = getattr(type_, '__origin__', None)
+    if origin is not None:
+        # Due to PEP560 we need to handle new typing
+        # classes carefully
+        return issubclass(type_.__origin__, Sequence)
+
     container = issubclass(type_, Sequence) or (issubclass(type_, Mapping))
     basic_scalars = issubclass(type_, str) or issubclass(type_, bytes)
 
-    return  (not basic_scalars) and container
+    return (not basic_scalars) and container
+
+
+def is_list_subclass(type_):
+    origin = getattr(type_, '__origin__', None)
+    if origin is not None:
+        return issubclass(type_.__origin__, list)
+    else:
+        return issubclass(type_, list)
 
 
 class File(FileStorage):
@@ -45,7 +58,7 @@ class Image(File):
     def __new__(cls, *args, **kwargs):
         try:
             from PIL import Image as pil_image
-        except ImportError as e:
+        except ImportError as e: # pragma: no cover
             e.args = ("Please install pillow to define an image type.",)
             raise e
         f = super().__new__(cls, *args, **kwargs)
@@ -60,49 +73,45 @@ class NDArray(File):
     def __new__(cls, *args, **kwargs):
         try:
             import numpy as np
-        except ImportError as e:
+        except ImportError as e: # pragma: no cover
             e.args = ("Please install NumPy to define an array type.",)
             raise e
         f = super().__new__(cls, *args, **kwargs)
         return np.load(f)
 
+def dt_factory(type_):
+    class ParsedDatetime(type):
+        '''A flexible datetime class
 
-class ParsedDateTime(Generic[T]):
-    '''A flexible dateteime.datetime class
+        receives a string returns datetime object
+        '''
+        __schema__ = {'type': 'string'}
+        __description__ = 'dateutil.parser.parse compatible datetime string'
+        def __new__(self, arg):
+            if type_ is datetime:
+                return parse(arg)
+            elif type_ is date:
+                return parse(arg).date()
+            else:
+                # most commonly used for pd.TimeStamp.
+                # any method that can take a string
+                return type_(arg)
 
-    receives a string: use dateutil to parse
-
-    The type specifier determines the returned type.
-    ParsedDateTime[datetime.date]
-    ParsedDateTime[datetime.datetime]
-    ParsedDateTime[pd.Timestamp]
-    '''
-    __schema__ = {'type':'string', 'format':'date-time'}
-    __description__ = 'dateutil.parser.parse compatible datetime string'
-
-    def __new__(cls, *args):
-        parsed =  parse(args[0])
-        _type = cls.__args__[0]
-        if issubclass(_type, date):
-            return parsed.date()
-        else:
-            return _type(parsed)
-        return parsed
+    return ParsedDatetime
 
 
-class TypedList(List, Generic[T]):
-    '''An dummy typed list
+def list_factory(type_):
+    items = getattr(type_, '__schema__', {}).get('type', PY_TYPES.get(type_, 'string'))
 
-    This class supports specialization with [].
-
-    fList = List[float]
-
-    It is expected to only receive one input.'''
-    __schema__ = {'type':'string'}
-
-    def __new__(cls, *args, **kwds):
-        _type = cls.__args__[0]
-        return _type(*args)
+    class TypedList(Sequence):
+        # using append with flask-restplus
+        # means that list is constructed later
+        __schema__ = {'type': items}
+        __description__ = 'List with values of type {}'.format(items)
+        def __new__(cls, arg):
+            return type_(arg)
+        
+    return TypedList
 
 
 def type_mapper(type_):
@@ -115,27 +124,29 @@ def type_mapper(type_):
     try:
         from PIL import Image as pil_image
         has_pil_image = True
-    except ImportError:
+    except ImportError: # pragma: no cover 
         has_pil_image = False
 
     try:
         import numpy as np
         has_numpy = True
-    except ImportError:
+    except ImportError: # pragma: no cover 
         has_numpy = False
-    
-    if issubclass(type_, List):
+
+    if is_list_subclass(type_):
         try:
             item_type = type_.__args__[0]
-        except:
+            if issubclass(item_type, (datetime, date)):
+                item_type = dt_factory(item_type)
+        except (AttributeError, TypeError):
             item_type = str
-        return TypedList[item_type]
+        return list_factory(item_type)
     elif issubclass(type_, TextIO):
         return TextFile
     elif issubclass(type_, BinaryIO):
         return File
     elif issubclass(type_, (datetime, date)):
-        return ParsedDateTime[type_]
+        return dt_factory(type_)
     elif has_pil_image and issubclass(type_, pil_image.Image):
         return Image
     elif has_numpy and issubclass(type_, np.ndarray):